+++ R/av.R
@@ -1,6 +1,6 @@
 ##
 ## internal
-## 
+##
 
 #' @importFrom httr status_code http_condition
 .avstop_for_status <-
@@ -20,6 +20,121 @@
     stop(message, call.=FALSE)
 }
 
+
+##
+## utilities
+##
+
+.avworkspace <- local({
+    hash <- new.env(parent = emptyenv())
+    function(fun, key, value) {
+        sysvar <- toupper(paste0("WORKSPACE_", key))
+        if (is.null(value)) {
+            if (is.null(hash[[key]])) {
+                ## initialize
+                hash[[key]] <- Sys.getenv(sysvar)
+                if (!nzchar(hash[[key]]) && interactive())
+                    warning("'", sysvar, "' undefined; use `", fun, "()` to set")
+            }
+        } else {
+            hash[[key]] <- ifelse(is.na(value), Sys.getenv(sysvar), value)
+        }
+        hash[[key]]
+    }
+})
+
+#' @rdname av
+#' @md
+#'
+#' @description `avworkspace_namespace()` and `avworkspace_name()` are
+#'     utiliity functions to retrieve workspace namespace and name
+#'     from environment variables or interfaces usually available in
+#'     AnVIL notebooks or RStudio sessions.  `avworkspace()` provides
+#'     a convenient way to specify workspace namespace and name in a
+#'     single command.
+#'
+#' @details `avworkspace_namespace()` is the billing account. If the
+#'     `namespace=` argument is not provided, try `gcloud_project()`,
+#'     and if that fails try `Sys.getenv("WORKSPACE_NAMESPACE")`.
+#'
+#'     `avworkspace_name()` is the name of the workspace as it appears
+#'     in \url{https://app.terra.bio/#workspaces}. If not provided,
+#'     `avworkspace_name()` tries to use
+#'     `Sys.getenv("WORKSPACE_NAME")`.
+#'
+#'     Values are cached across sessions, so explicitly providing
+#'     `avworkspace_*()` is required at most once per session. Revert
+#'     to system settings with arguments `NA`.
+#'
+#' @param namespace character(1) AnVIL workspace namespace as returned
+#'     by, e.g., `avworkspace_namespace()`
+#'
+#' @param name character(1) AnVIL workspace name as returned by, eg.,
+#'     `avworkspace_name()`.
+#'
+#' @return `avworkspace_namespace()`, and `avworkspace_name()` return
+#'     `character(1)` identifiers.
+#'
+#' @examples
+#' avworkspace_namespace()
+#'
+#' @export
+avworkspace_namespace <- function(namespace = NULL) {
+    suppressWarnings({
+        namespace <- .avworkspace("avworkspace_namespace", "NAMESPACE", namespace)
+    })
+    if (!nzchar(namespace)) {
+        namespace <- tryCatch({
+            gcloud_project()
+        }, error = function(e) {
+            NULL
+        })
+        namespace <- .avworkspace("avworkspace_namespace", "NAMESPACE", namespace)
+    }
+    namespace
+}
+
+#' @rdname av
+#'
+#' @examples
+#' avworkspace_name()
+#'
+#' @export
+avworkspace_name <- function(name = NULL)
+    .avworkspace("avworkspace_name", "NAME", name)
+
+#' @rdname av
+#'
+#' @param workspace when present, a `character(1)` providing the
+#'     concatenated namespace and name, e.g.,
+#'     `"bioconductor-rpci-anvil/Bioconductor-Package-AnVIL"`
+#'
+#' @return `avworkspace()` returns the character(1) concatenated
+#'     namespace and name.
+#'
+#' @examples
+#' avworkspace()
+#'
+#' @export
+avworkspace <-
+    function(workspace = NULL)
+{
+    stopifnot(
+        is.null(workspace) || .is_scalar_character(workspace)
+    )
+    if (!is.null(workspace)) {
+        wkspc <- strsplit(workspace, "/")[[1]]
+        if (length(wkspc) != 2L)
+            stop(
+                "'workspace' must be of the form 'namespace/name', ",
+                "with a single '/'"
+            )
+        avworkspace_namespace(wkspc[[1]])
+        avworkspace_name(wkspc[[2]])
+    }
+    paste0(avworkspace_namespace(), "/", avworkspace_name())
+}
+
 ##
 ## tables
 ##
@@ -30,7 +145,11 @@
 #'
 #' @description `avtables()` describes tables available in a
 #'     workspace. Tables can be visualized under the DATA tab, TABLES
-#'     item.
+#'     item.  `avtable()` returns an AnVIL table.  `avtable_import()`
+#'     imports a data.frame to an AnVIL table.  `avtable_import_set()`
+#'     imports set membership (i.e., a subset of an existing table)
+#'     information to an AnVIL table.  `avtable_delete_values()`
+#'     removes rows from an AnVIL table.
 #'
 #' @return `avtables()`: A tibble with columns identifying the table,
 #'     the number of records, and the column names.
@@ -59,8 +178,6 @@ avtables <-
 }
 
 #' @rdname av
-#'
-#' @description `avtable()` returns an AnVIL table.
 #'
 #' @param table character(1) table name as returned by, e.g., `avtables()`.
 #'
@@ -114,8 +231,6 @@ avtable <-
 
 #' @rdname av
 #'
-#' @description `avtable_import()` imports a data.frame to an AnVIL table.
-#'
 #' @param .data A tibble or data.frame for import as an AnVIL table.
 #'
 #' @param entity `character(1)` column name of `.data` to be used as
@@ -162,7 +277,78 @@ avtable_import <-
 
 #' @rdname av
 #'
-#' @description `avtable_delete_values()` removes rows from an AnVIL table.
+#' @param origin character(1) name of the entity (table) used to
+#'     create the set e.g "sample", "participant",
+#'     etc.
+#'
+#' @param set `character(1)` column name of `.data` identifying the
+#'     set(s) to be created.
+#'
+#' @param member `character()` vector of entity from the avtable
+#'     identified by `origin`. The values may repeat if an ID is in
+#'     more than one set
+#'
+#' @details `avtable_import_set()` creates new rows in a table
+#'     `<origin>_set`. One row will be created for each distinct value
+#'     in the column identified by `set`. Each row entry has a
+#'     corresponding column `<origin>` linking to one or more rows in
+#'     the `<origin>` table, as given in the `member` column. The
+#'     operation is somewhat like `split(member, set)`.
+#'
+#' @return `avtable_import_set()` returns a `character(1)` name of the
+#'     imported AnVIL tibble.
+#'
+#' @importFrom utils write.table
+#'
+#' @examples
+#' \dontrun{
+#' ## editable copy of '1000G-high-coverage-2019' workspace
+#' avworkspace("bioconductor-rpci-anvil/1000G-high-coverage-2019")
+#' sample <-
+#'     avtable("sample") %>%                               # existing table
+#'     mutate(set = sample(head(LETTERS), nrow(.), TRUE))  # arbitrary groups
+#' sample %>%                                   # new 'participant_set' table
+#'     avtable_import_set("participant", "set", "participant")
+#' sample %>%                                   # new 'sample_set' table
+#'     avtable_import_set("sample", "set", "name")
+#' }
+#'
+#' @export
+avtable_import_set <-
+    function(
+        .data, origin, set = names(.data)[[1]], member = names(.data)[[2]],
+        namespace = avworkspace_namespace(), name = avworkspace_name())
+{
+    stopifnot(
+        is.data.frame(.data),
+        .is_scalar_character(origin),
+        .is_scalar_character(set),
+        set %in% names(.data),
+        .is_scalar_character(member),
+        !identical(set, member), member %in% names(.data),
+        .is_scalar_character(namespace),
+        .is_scalar_character(name)
+    )
+
+    origin <- curl_escape(origin)
+
+    tbl <-
+        .data %>%
+        select(set, member)
+    names(tbl)[[1]] <- paste0("membership:", origin, "_set_id")
+    names(tbl)[[2]] <- origin
+
+    destination <- tempfile()
+    write.table(tbl, destination, quote = FALSE, sep="\t", row.names=FALSE)
+
+    entities <- httr::upload_file(destination)
+    response <- Terra()$flexibleImportEntities(namespace, name, entities)
+    .avstop_for_status(response, "avtable_import_sample_set")
+
+    httr::content(response, type="text", encoding = "UTF-8")
+}
+
+#' @rdname av
 #'
 #' @param values vector of values in the entity (key) column of
 #'     `table` to be deleted. A table `sample` has an associated
@@ -208,7 +394,7 @@ avtable_delete_values <-
 #'     bucket) of the element.
 #'
 #' @examples
-#' if (gcloud_exists() && nzchar(avworkspace_namespace()))
+#' if (gcloud_exists() && nzchar(avworkspace_name()))
 #'     ## from within AnVIL
 #'     avdata()
 #'
@@ -278,7 +464,7 @@ avdata <-
 #'     prefixed with `gs://` if `as_path = TRUE`.
 #'
 #' @examples
-#' if (gcloud_exists() && nzchar(avworkspace_namespace()))
+#' if (gcloud_exists() && nzchar(avworkspace_name()))
 #'     ## From within AnVIL...
 #'     bucket <- avbucket()                        # discover bucket
 #'
@@ -341,7 +527,11 @@ avbucket <-
 #' @rdname av
 #'
 #' @description `avfiles_ls()` returns the paths of files in the
-#'     workspace bucket.
+#'     workspace bucket.  `avfiles_backup()` copies files from the
+#'     compute node file system to the workspace bucket.
+#'     `avfiles_restore()` copies files from the workspace bucket to
+#'     the compute node file system.  `avfiles_rm()` removes files or
+#'     directories from the workspace bucket.
 #'
 #' @param path For `avfiles_ls(), the character(1) file or directory path
 #'     to list.
@@ -364,7 +554,7 @@ avbucket <-
 #'     workspace bucket.
 #'
 #' @examples
-#' if (gcloud_exists() && nzchar(avworkspace_namespace()))
+#' if (gcloud_exists() && nzchar(avworkspace_name()))
 #'     avfiles_ls()
 #'
 #' @export
@@ -373,7 +563,7 @@ avfiles_ls <-
         path = "",
         full_names = FALSE,
         recursive = FALSE,
-        namespace = avworkspace_namespace(), 
+        namespace = avworkspace_namespace(),
         name = avworkspace_name())
 {
     stopifnot(
@@ -393,9 +583,6 @@ avfiles_ls <-
 }
 
 #' @rdname av
-#'
-#' @description `avfiles_backup()` copies files from the compute node
-#'     file system to the workspace bucket.
 #'
 #' @details `avfiles_backup()` can be used to back-up individual files
 #'     or entire directories, recursively.  When `recursive = FALSE`,
@@ -457,9 +644,6 @@ avfiles_backup <-
 
 #' @rdname av
 #'
-#' @description `avfiles_restore()` copies files from the workspace
-#'     bucket to the compute node file system.
-#'
 #' @details `avfiles_restore()` behaves in a manner analogous to
 #'     `avfiles_backup()`, copying files from the workspace bucket to
 #'     the compute node file system.
@@ -492,9 +676,6 @@ avfiles_restore <-
 
 #' @rdname av
 #'
-#' @description `avfiles_rm()` removes files or directories from the
-#'     workspace bucket.
-#'
 #' @return `avfiles_rm()` on success, returns a list of the return
 #'     codes of `gsutil_rm()`, invisibly.
 #'
@@ -506,7 +687,7 @@ avfiles_rm <-
         parallel = TRUE,
         namespace = avworkspace_namespace(),
         name = avworkspace_name()
-    )        
+    )
 {
     stopifnot(
         .is_character(source),
@@ -525,62 +706,152 @@ avfiles_rm <-
 }
 
 ##
-## utilities
+## runtimes / persistent disks
 ##
 
-.avworkspace <- local({
-    hash <- new.env(parent = emptyenv())
-    function(fun, key, value) {
-        sysvar <- toupper(paste0("WORKSPACE_", key))
-        if (is.null(value)) {
-            if (is.null(hash[[key]])) {
-                ## initialize
-                hash[[key]] <- Sys.getenv(sysvar)
-                if (!nzchar(hash[[key]]) && interactive())
-                    warning("'", sysvar, "' undefined; use `", fun, "()` to set")
-            }
-        } else {
-            hash[[key]] <- ifelse(is.na(value), Sys.getenv(sysvar), value)
-        }
-        hash[[key]]
-    }
-})
-
 #' @rdname av
+#' @md
 #'
-#' @description `avworkspace_namespace()` and `avworkspace_name()` are
-#'     utiliity functions to retrieve workspace namespace and name
-#'     from environment variables or interfaces usually available in
-#'     AnVIL notebooks or RStudio sessions.
+#' @description `avruntimes()` returns a tibble containing information
+#'     about runtimes (notebooks or RStudio instances, for example)
+#'     that the current user has access to.
 #'
-#' @details `avworkspace_namespace()` is usually the billing account,
-#'     and `avworkspace_name()` the name of the workspace as it
-#'     appears in \url{https://app.terra.bio/#workspaces}. Providing
-#'     arguments to these functions over-rides AnVIL-determined
-#'     settings with the provided value. Revert to system settings
-#'     with arguments `NA`.
+#' @return `avruntimes()` returns a tibble with columns
 #'
-#' @param namespace character(1) AnVIL workspace namespace as returned
-#'     by, e.g., `avworkspace_namespace()`
-#'
-#' @param name character(1) AnVIL workspace name as returned by, eg.,
-#'     `avworkspace_name()`.
-#'
-#' @return `avworkspace_namespace()`, and `avworkspace_name()` return
-#'     `character(1)` identifiers.
+#' - id: integer() runtime identifier.
+#' - googleProject: character() billing account.
+#' - tool: character() e.g., "Jupyter", "RStudio".
+#' - status character() e.g., "Stopped", "Running".
+#' - creator character() AnVIL account, typically "user@gmail.com".
+#' - createdDate character() creation date.
+#' - destroyedDate character() destruction date, or NA.
+#' - dateAccessed character() date of (first?) access.
+#' - runtimeName character().
+#' - clusterServiceAccount character() service ('pet') account for
+#'   this runtime.
+#' - masterMachineType character() It is unclear which 'tool' populates
+#'   which of the machineType columns).
+#' - workerMachineType character().
+#' - machineType character().
+#' - persistentDiskId integer() identifier of persistent disk (see
+#'   `avdisks()`), or `NA`.
 #'
 #' @examples
-#' avworkspace_namespace()
+#' if (gcloud_exists())
+#'     ## from within AnVIL
+#'     avruntimes()
+#'
+#' @importFrom dplyr rename_with
+#'
+#' @importFrom tidyselect everything
 #'
 #' @export
-avworkspace_namespace <- function(namespace = NULL)
-    .avworkspace("avworkspace_namespace", "NAMESPACE", namespace)
+avruntimes <-
+    function()
+{
+    template <- list(
+        id = integer(0),
+        googleProject = character(0),
+        labels.tool = character(0),
+        status = character(0),
+        auditInfo.creator = character(0),
+        auditInfo.createdDate = character(0),
+        auditInfo.destroyedDate = logical(0),
+        auditInfo.dateAccessed = character(0),
+        runtimeName = character(0),
+        labels.clusterServiceAccount = character(0),
+        runtimeConfig.masterMachineType = character(0),
+        runtimeConfig.workerMachineType = logical(0),
+        runtimeConfig.machineType = character(0),
+        runtimeConfig.persistentDiskId = integer(0)
+    )
 
-#' @rdname av
+    leo <- Leonardo()
+    response <- leo$listRuntimes()
+    .avstop_for_status(response, "avruntimes")
+    runtimes <- flatten(response)
+
+    .tbl_with_template(runtimes, template) %>%
+        rename_with(~ sub(".*\\.", "", .x))
+}
+
+#' @importFrom dplyr pull
+.runtime_pet <-
+    function(creator, tool = c("Jupyter", "RStudio"),
+             namespace = avworkspace_namespace())
+{
+    tool <- match.arg(tool)
+    stopifnot(
+        .is_scalar_character(tool),
+        .is_scalar_character(creator),
+        .is_scalar_character(namespace)
+    )
+
+    runtimes <- avruntimes()
+    pet <-
+        runtimes %>%
+        filter(
+            .data$tool == {{ tool }},
+            .data$creator == {{ creator }},
+            .data$googleProject == {{ namespace }}
+        ) %>%
+        pull(.data$clusterServiceAccount)
+
+    if (!.is_scalar_character(pet))
+        warning("'.runtime_pet' return value is not scalar")
+    pet
+}
+
+#' @name av
+#' @md
+#'
+#' @description 'avdisks()` returns a tibble containing information
+#'     about persistent disks associatd with the current user.
+#'
+#' @return `avdisks()` returns a tibble with columns
+#'
+#' - id character() disk identifier.
+#' - googleProject: character() billing account.
+#' - status, e.g, "Ready"
+#' - size integer() in GB.
+#' - diskType character().
+#' - blockSize integer().
+#' - creator character() AnVIL account, typically "user@gmail.com".
+#' - createdDate character() creation date.
+#' - destroyedDate character() destruction date, or NA.
+#' - dateAccessed character() date of (first?) access.
+#' - zone character() e.g.. "us-central1-a".
+#' - name character().
 #'
 #' @examples
-#' avworkspace_name()
+#' if (gcloud_exists())
+#'     ## from within AnVIL
+#'     avdisks()
 #'
 #' @export
-avworkspace_name <- function(name = NULL)
-    .avworkspace("avworkspace_name", "NAME", name)
+avdisks <-
+    function()
+{
+    template <- list(
+        id = integer(0),
+        googleProject = character(0),
+        status = character(0),
+        size = integer(0),
+        diskType = character(0),
+        blockSize = integer(0),
+        auditInfo.creator = character(0),
+        auditInfo.createdDate = character(0),
+        auditInfo.destroyedDate = logical(0),
+        auditInfo.dateAccessed = character(0),
+        name = character(0),
+        zone = character(0)
+    )
+
+    leo <- Leonardo()
+    response <- leo$listDisks()
+    .avstop_for_status(response, "avdisks")
+    runtimes <- flatten(response)
+
+    .tbl_with_template(runtimes, template) %>%
+        rename_with(~sub(".*\\.", "", .x))
+}