+++ NEWS.md
@@ -12,6 +12,18 @@ NEW FEATURES
 
 - (v 1.1.9) add avfiles_*() for managing workspace bucket files.
 
+- (v 1.1.15) add avtable_import_set() to create subsets of tables,
+  following the Terra data model.
+
+- (v 1.1.16) add avruntimes(), avworkspace_jobs() to query for runtimes
+  and jobs associated with the active billing account.
+
+- (v 1.1.17) add avdisks() to query for persistent disks associate
+  with the active billing account.
+  
+- (v 1.1.21) add avworkflow_*() for interacting with workflow jobs
+  and outputs.
+
 # AnVIL 1.0.x
 
 BUG FIXES
@@ -27,8 +39,7 @@ USER VISIBLE CHANGES
 
 # AnVIL 1.0.0
 
-- AnVIL is _finally
-_ on Bioconductor!
+- AnVIL is _finally on Bioconductor!
 - Support OpenAPI Specification version 2 (aka Swagger 2.0)
 - `av`, `gcloud`, `gsutil` type functions added to interface with AnVIL
   and the cloud